requirements.txt
PyYAML==6.0.1
scrapli==2023.7.30